Added code to provision firewall rule. Consolidated build service functions. (#181)
* Added code to provision firewall. Consolidated build service functions.

* Added python lint pre-commit hook.

* testing pre-commit python lint

* testing pre-commit hook

* testing pre-commit hook

* testing pre-commit hook

* testing pre-commit hook

* testing pre-commit hook

* testing pre-commit hook

* testing pre-commit hook

* Broke up gcp_api_util module into cloudsql, compute and storage modules. Improved comments

* Updated tests to work with no module structure. Corrected spacing in compute module.

* Fixed comments.

* Fixed comments.
